+++ src/global.d.ts
@@ -1,0 +1,13 @@
+export namespace ENV {
+  interface ProcessEnv {
+    STANDARD_CONNECTION: string;
+    NODE_ENV: string;
+    DB_HOST: string;
+    DB_USER: string;
+    DB_PWD: string;
+    DB_PORT: string;
+    DB_NAME: string;
+    ACCTOKEN_LIFE: string,
+    ACCTOKEN_SECRET: string
+  }
+}